Reworked Xspec interface, added C support
